requirements.txt
fastapi
firebase-admin
python-telegram-bot
python-decouple
pydantic

pytest